+++ ChangeLog.md
@@ -4,6 +4,9 @@ ChangeLog {#changelog}
 # For next stable IMP release #
 - The IMP::base and IMP::kernel namespaces have been removed (use the top-level
   IMP namespace instead).
+- IMP::Optimizer and IMP::Sampler now require an IMP::ScoringFunction to be set
+  before they are used (since IMP::Model no longer has an implicit scoring
+  function).
 
 # 2.5.0 - 2015-06-30 # {#changelog_2_5_0}
 - Documentation has been revised and updated; the main manual now gives a